requirements.txt
aiogram==3.4.1
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
aiofiles==23.2.1
websockets==12.0 